reading parquet or delta files from local directory or aws s3 or azure blob

reads parquet,delta files from local or cloud

readparquetR(
  pathtoread,
  where = "",
  partition = "",
  collist = "",
  sample = F,
  samplesizecount = 3,
  add_part_names = F,
  format = "parquet",
  filelocation = "local",
  containerconnection = NULL,
  bucket = NULL
)

Arguments

pathtoread

reading path, local or azure cloud

where

it will read datatable and filter with this condition. i.e. you can write where="column='A'"

partition

if you want to read a partition files with a pattern. i.e. partition=c('2017','2018')

collist

specific columns to read

sample

sample=T just to see sample rows. dont read whole table

samplesizecount

default=3 rows. you can change it

add_part_names

when it is partitioned, you need to make this T to add partition names as column

filelocation

"local" or "azure" or "s3"

containerconnection

if filelocation="azure" then we need connection name

bucket

if filelocation="s3" we need to put bucket name

Examples

readparquetR(pathtoread="localfolder/", collist = c("column1","column2","column3"),format="delta",where="SKU==19058901 & STORE!='1905'")
#> [1] "localfolder/"
#> [1] "there is no row, make sure path is correct!"
#> Error in .checkTypos(e, names_x): Object 'x' not found amongst 
readparquetR(pathtoread="C:/users/...",add_part_names=F,collist="",sample=F,where="sku=1 & store=1",partition="2022")
#> [1] "C:/users/..."
#> [1] "there is no row, make sure path is correct!"
#> Null data.table (0 rows and 0 cols)
your_connection = AzureStor::storage_container(AzureStor::storage_endpoint(your_link, key=your_key), "your_container")
#> Error in is_url(url): object 'your_link' not found